feat: move SongItem to search folder
