+++ frontend/src/search/components/SongItem/SongItem.types.ts
@@ -1,0 +1,8 @@
+import { Song } from "@chordium/types";
+
+export interface SongItemProps {
+  item: Song;
+  onView: (song: Song) => void;
+  onDelete: (songId: string) => void;
+  style?: React.CSSProperties;
+}
