+++ frontend/src/search/components/SongItem/index.ts
@@ -1,0 +1,1 @@
+export { default as SongItem } from "./SongItem";
