+++ frontend/src/search/components/index.ts
@@ -1,3 +1,4 @@
 export * from "./SearchResults";
 export * from "./SearchTab";
 export * from "./ArtistResults";
+export * from "./SongItem";
